Progress bar for warc injector.

+++ script/inject/__main__.py
@@ -11,14 +11,14 @@ import sqlite3
 import datetime
 import sys
 import time
-import flask
+# import flask
 import signal, os
 import random
 from itertools import repeat
 staleTime = datetime.timedelta(7,0,0) # one week for now
 
-app = flask.Flask(__name__)
-app.secret_key = 'oaisj84alwsdkjhf9238u'
+# app = flask.Flask(__name__)
+# app.secret_key = 'oaisj84alwsdkjhf9238u'
 
 def getDb(makeDates=True):
     if makeDates:
@@ -106,13 +106,19 @@ def injectItem(item, db, mode):
 
     dbUpdates = []
     skipped = 0
-    for ff in md['files']:
-        if not ff['name'].endswith('arc.gz'): continue
+    warcs = filter(lambda x: 'name' in x and x['name'].endswith and x['name'].endswith('arc.gz'), md['files'])
+
+    collectionName = metadata.get('archiveit-collection-name', '')
+    for ii, ff in enumerate(warcs):
+        #if not ff['name'].endswith('arc.gz'): continue
         itemMetadata = {'mtime':ff['mtime']}
         updateTime = datetime.datetime.fromtimestamp(float(ff['mtime']))
         if mode != 'force' and ff['name'] in lastUpdate and updateTime <= lastUpdate[ff['name']]:
             print "skip {0} because it is up to date".format(ff['name'])
             skipped += 1
+            requests.post('http://localhost:10008/progress', 
+                          json={'item':item, 'total':len(warcs), 'done':ii+1, 
+                                'collection-name':collectionName})
             continue
         
         itemMetadata.update(md['metadata'])
@@ -140,18 +146,24 @@ def injectItem(item, db, mode):
         sys.stdout.flush()
             
         dbUpdates.append((itemId, ff['name'], updateTime, statusCode, took))
+        requests.post('http://localhost:10008/progress', 
+                      json={'item':item, 'total':len(warcs), 'done':ii+1, 
+                            'collection-name':collectionName})
 
-    reallyExecuteMany(c, "DELETE FROM files where fileName = ? ", zip(lastUpdate.iterkeys()))
-    reallyExecuteMany(c, "INSERT INTO files VALUES (?,?,?,?,?)",
-                      dbUpdates)
-    db.commit()
 
+    if len(dbUpdates):
+        reallyExecuteMany(c, "DELETE FROM files where fileName = ? ", zip(lastUpdate.iterkeys()))
+        reallyExecuteMany(c, "INSERT INTO files VALUES (?,?,?,?,?)",
+                          dbUpdates)
+        db.commit()
     print 'completed %s with %s items injected and %s skipped' % (item, len(dbUpdates), skipped)
 
+
 def getPage(zippedArgs):
-    page, mode = zippedArgs
+    page, mode, resultsPerPage, extraQuery = zippedArgs
+    query = 'collection%3Aarchiveitdigitalcollection+' + extraQuery
     #r = requests.get('https://archive.org/advancedsearch.php?q=collection%3Aarchiveitdigitalcollection&fl%5B%5D=identifier&rows=1&page={0}&output=json&save=yes'.format(page))
-    r = requests.get('https://archive.org/advancedsearch.php?q=collection%3Aarchiveitdigitalcollection&fl%5B%5D=identifier&sort[]=date+desc&rows=100&page={0}&output=json&save=yes'.format(page))
+    r = requests.get('https://archive.org/advancedsearch.php?q={1}&fl%5B%5D=identifier&sort[]=date+asc&rows={2}&page={0}&output=json'.format(page, query, resultsPerPage))
     if r.status_code != 200:
         return 0
 
@@ -161,6 +173,7 @@ def getPage(zippedArgs):
     numFound = jsonContents['response']['numFound']
                   
     if len(items) == 0:
+        requests.post('http://localhost:10008/progress', json={'total':numFound, 'completed':'', 'query':extraQuery})
         print 'got 0 items for search page', page
         return 0
     print 'loading %s items, %s - %s of %s' % (len(items), items[0], items[-1], numFound)
@@ -169,7 +182,7 @@ def getPage(zippedArgs):
 
     for item in items:
         injectItem(item, db, mode)
-
+        requests.post('http://localhost:10008/progress', json={'total':numFound, 'completed':item, 'query':extraQuery})
     db.close()
     return len(items)
 
@@ -199,6 +212,10 @@ def showItems():
 
 
 def nuke(lastPid, fromOrbit=False):
+    try:
+        requests.post('http://localhost:10008/shutdown', {})
+    except:
+        pass
     sig = signal.SIGTERM
     if fromOrbit:
         sig = signal.SIGKILL
@@ -222,16 +239,46 @@ def nuke(lastPid, fromOrbit=False):
 
 def main():
     global staleTime
-    try:
-        lastPid = open('running.pid', 'r').read()
-    except:
-        lastPid = None
     print 'arguments were', sys.argv, 'pid is', os.getpid()
-    open('running.pid', 'w').write(str(os.getpid()))
+
+    if sys.argv[1] != 'monitor':
+        try:
+            lastPid = open('running.pid', 'r').read()
+        except:
+            lastPid = None
+        open('running.pid', 'w').write(str(os.getpid()))
 
     # p = multiprocessing.Process(target=serveForever)
     # p.start()
+    
+    if sys.argv[1] == 'test':
+        query = ''
+        if len(sys.argv) == 3:
+            query = sys.argv[2]
+
+        subprocess.Popen(['python','inject', 'monitor'])
+        
+        mode = 'testing'
+        runInjects(10, 'testing', query)
+
+    if sys.argv[1] == 'run':
+        query = ''
+        if len(sys.argv) == 4:
+            query = sys.argv[3]
+
+            subprocess.Popen(['./warc-inject','monitor'])
+        threads = int(sys.argv[2])
+        runInjects(threads, 'production', query)
+        print "done running"
+
+
+
+
     if len(sys.argv) == 2:
+        if sys.argv[1] == 'monitor':
+            import monitor
+            monitor.main()
+
         if sys.argv[1] == 'init':
             init()
             print 'initialized'
@@ -253,6 +300,8 @@ def main():
             nuke(lastPid, fromOrbit=True)
 
         if sys.argv[1] == 'test':
+            subprocess.Popen(['./warc-inject','monitor'])
+
             mode = 'testing'
             runInjects(10, 'testing')
 
@@ -314,8 +363,8 @@ def main():
 
             signal.alarm(0)          # Disable the alarm
 
-        if sys.argv[1] == 'serve':
-            serveForever()
+        # if sys.argv[1] == 'serve':
+        #     serveForever()
 
     if len(sys.argv) == 3:
         if sys.argv[1] == 'force':
@@ -324,10 +373,6 @@ def main():
             injectItem(itemName, db, 'production')
             sys.exit(0)
 
-        if sys.argv[1] == 'run':
-            threads = int(sys.argv[2])
-            runInjects(threads)
-            print "done running"
 
 
     if len(sys.argv) == 4:
@@ -383,19 +428,41 @@ def main():
             sys.exit(0)
 
 
+def getNumResults(query):
+    query = 'collection%3Aarchiveitdigitalcollection+' + query
+    r = requests.get('https://archive.org/advancedsearch.php?q={0}&fl%5B%5D=identifier&sort[]=date+asc&rows=1&page=0&output=json'.format(query))
+    if r.status_code != 200:
+        return 0
+    contents = r.content
+    jsonContents = json.loads(contents)
+    numFound = jsonContents['response']['numFound']
+    return numFound
+    
 
 
         
-def runInjects(threads, mode='production'):
+def runInjects(threads, mode='production', query=''):
     from multiprocessing.pool import ThreadPool
+    import math
     pool = ThreadPool(processes=threads)
     try:
-        maxPages = 1300
-        answer = pool.map(getPage, zip(xrange(1,maxPages), repeat(mode, maxPages)))
+        totalResults = getNumResults(query)
+        resultsPerPage = 10
+        maxPages = int(math.ceil(totalResults / float(resultsPerPage)))
+        if maxPages < threads:
+            maxPages = threads
+            resultsPerPage = int(math.ceil(totalResults / float(maxPages)))
+        print threads, ' threads,', totalResults, 'total,', maxPages, 'pages', resultsPerPage, 'results per page'
+        answer = pool.map(getPage, zip(xrange(1,maxPages), 
+                                       repeat(mode, maxPages),
+                                       repeat(resultsPerPage, maxPages),
+                                       repeat(query, maxPages)))
         print "finished item pass", answer
     except (KeyboardInterrupt, SystemExit):
         print 'ok, caught'
-        raise
+        requests.post('http://localhost:10008/shutdown', {})
+        sys.exit(0)
+        #raise
 
 
 def init():
@@ -412,67 +479,67 @@ def init():
     db.close()
 
 
-def serveForever():
-    @app.route('/',
-               methods=['GET', 'POST'], endpoint='home')
-    def home():
-        db = getDb(makeDates=False)
-        res = db.execute('select * from items limit 10')
-        for item, checked in res.fetchall():
-            print item
-            try:
-                metadata = subprocess.Popen(['./ia','metadata', item],
-                                            stdout=subprocess.PIPE).communicate()[0]
+# def serveForever():
+#     @app.route('/',
+#                methods=['GET', 'POST'], endpoint='home')
+#     def home():
+#         db = getDb(makeDates=False)
+#         res = db.execute('select * from items limit 10')
+#         for item, checked in res.fetchall():
+#             print item
+#             try:
+#                 metadata = subprocess.Popen(['./ia','metadata', item],
+#                                             stdout=subprocess.PIPE).communicate()[0]
 
-                break
-            except Exception, e:
-                pass
-        db.close()
+#                 break
+#             except Exception, e:
+#                 pass
+#         db.close()
 
-        return flask.make_response(metadata)
+#         return flask.make_response(metadata)
 
-    @app.route('/progress',
-               methods=['GET', 'POST'], endpoint='progress')
-    def progress():
-        r = requests.get('https://archive.org/advancedsearch.php?q=collection%3Aarchiveitdigitalcollection&fl%5B%5D=identifier&sort[]=date+desc&rows=1&page=1&output=json')
-        if r.status_code != 200:
-            return flask.make_response(json.dumps({error:'ia search feed is down'}), 
-                                       'application/json')
+#     @app.route('/progress',
+#                methods=['GET', 'POST'], endpoint='progress')
+#     def progress():
+#         r = requests.get('https://archive.org/advancedsearch.php?q=collection%3Aarchiveitdigitalcollection&fl%5B%5D=identifier&sort[]=date+desc&rows=1&page=1&output=json')
+#         if r.status_code != 200:
+#             return flask.make_response(json.dumps({error:'ia search feed is down'}), 
+#                                        'application/json')
 
-        contents = r.content
-        jsonContents = json.loads(contents)
-        numFound = jsonContents['response']['numFound']
-
-
-        db = getDb()
-        examinedItems = db.execute('select count(*) from items').fetchone()
-        itemsWithWarc = db.execute('select count(*) from items where ROWID in (select itemId from files where files.status = 200)').fetchone()
-        return flask.make_response(json.dumps({'totalItems':numFound, 
-                                               'examinedItems':examinedItems,
-                                               'itemsWithWarc':itemsWithWarc
-                                           }, indent=4), 'application/json')
+#         contents = r.content
+#         jsonContents = json.loads(contents)
+#         numFound = jsonContents['response']['numFound']
 
 
-    @app.route('/items',
-               methods=['GET', 'POST'], endpoint='items')
-    def items():
-        db = getDb(makeDates=False)
+#         db = getDb()
+#         examinedItems = db.execute('select count(*) from items').fetchone()
+#         itemsWithWarc = db.execute('select count(*) from items where ROWID in (select itemId from files where files.status = 200)').fetchone()
+#         return flask.make_response(json.dumps({'totalItems':numFound, 
+#                                                'examinedItems':examinedItems,
+#                                                'itemsWithWarc':itemsWithWarc
+#                                            }, indent=4), 'application/json')
 
-        c = db.cursor()
-        res = c.execute("select item, checked from items")    
 
-        out = []
-        for item, checked in res.fetchall():
-            out.append({'item':item, 'checked':checked})
-        db.close()
+#     @app.route('/items',
+#                methods=['GET', 'POST'], endpoint='items')
+#     def items():
+#         db = getDb(makeDates=False)
 
-        return flask.make_response(json.dumps(out), 'application/json')
+#         c = db.cursor()
+#         res = c.execute("select item, checked from items")    
 
-    app.run('0.0.0.0', 
-            port=7999,
-            debug=False,
-            use_reloader=False,
-            use_debugger=False)
+#         out = []
+#         for item, checked in res.fetchall():
+#             out.append({'item':item, 'checked':checked})
+#         db.close()
+
+#         return flask.make_response(json.dumps(out), 'application/json')
+
+#     app.run('0.0.0.0', 
+#             port=7999,
+#             debug=False,
+#             use_reloader=False,
+#             use_debugger=False)
 
 
 if __name__ == '__main__':
